feat: display existing validation checks in publish dialog (#3276)
