+++ editor.planx.uk/src/pages/FlowEditor/components/Sidebar/index.tsx
@@ -7,33 +7,27 @@ import SignalCellularAltIcon from "@mui/icons-material/SignalCellularAlt";
 import Badge from "@mui/material/Badge";
 import Box from "@mui/material/Box";
 import Button from "@mui/material/Button";
-import Collapse from "@mui/material/Collapse";
 import Container from "@mui/material/Container";
 import Dialog from "@mui/material/Dialog";
 import DialogActions from "@mui/material/DialogActions";
 import DialogContent from "@mui/material/DialogContent";
 import DialogTitle from "@mui/material/DialogTitle";
-import Divider from "@mui/material/Divider";
 import Link from "@mui/material/Link";
-import List from "@mui/material/List";
-import ListItem from "@mui/material/ListItem";
-import ListItemText from "@mui/material/ListItemText";
-import { styled } from "@mui/material/styles";
 import Tab, { tabClasses } from "@mui/material/Tab";
 import Tabs from "@mui/material/Tabs";
 import Tooltip from "@mui/material/Tooltip";
 import Typography from "@mui/material/Typography";
-import { ComponentType as TYPES } from "@opensystemslab/planx-core/types";
+import { styled } from "@mui/material/styles";
 import { AxiosError } from "axios";
 import React, { useState } from "react";
 import { useAsync } from "react-use";
-import Caret from "ui/icons/Caret";
 import Input from "ui/shared/Input";
 
 import { formatLastPublishMessage } from "pages/FlowEditor/utils";
 import Questions from "../../../Preview/Questions";
 import { useStore } from "../../lib/store";
 import EditHistory from "./EditHistory";
+import { AlteredNode, AlteredNodesSummaryContent, ValidationCheck, ValidationChecks } from "./PublishDialog";
 
 type SidebarTabs = "PreviewBrowser" | "History";
 
@@ -151,211 +145,6 @@ const DebugConsole = () => {
   );
 };
 
-interface AlteredNode {
-  id: string;
-  type: TYPES;
-  data?: any;
-}
-
-const AlteredNodeListItem = (props: { node: AlteredNode }) => {
-  const { node } = props;
-  let text, data;
-
-  if (node.id === "_root") {
-    text = "Changed _root service by adding, deleting or re-ordering nodes";
-  } else if (node.id === "0") {
-    text = `The entire _root service will be published for the first time`;
-  } else if (node.id && Object.keys(node).length === 1) {
-    text = `Deleted node ${node.id}`;
-  } else if (node.type && node.data) {
-    text = `Added/edited ${TYPES[node.type]}`;
-    data = JSON.stringify(node.data, null, 2);
-  } else {
-    text = `Added/edited ${TYPES[node.type]}`;
-  }
-
-  return (
-    <li key={node.id}>
-      <Typography variant="body2">{text}</Typography>
-      {data && <pre style={{ fontSize: ".8em" }}>{data}</pre>}
-    </li>
-  );
-};
-
-interface Portal {
-  text: string;
-  flowId: string;
-  publishedFlowId: number;
-  summary: string;
-  publishedBy: number;
-  publishedAt: string;
-}
-
-const AlteredNestedFlowListItem = (props: Portal) => {
-  const { text, flowId, publishedFlowId, summary, publishedAt } = props;
-
-  const [nestedFlowLastPublishedTitle, setNestedFlowLastPublishedTitle] =
-    useState<string>();
-  const lastPublisher = useStore((state) => state.lastPublisher);
-
-  const _nestedFlowLastPublishedRequest = useAsync(async () => {
-    const user = await lastPublisher(flowId);
-    setNestedFlowLastPublishedTitle(formatLastPublishMessage(publishedAt, user));
-  });
-
-  return (
-    <ListItem
-      key={publishedFlowId}
-      disablePadding
-      sx={{ display: "list-item" }}
-    >
-      <ListItemText
-        primary={
-          useStore.getState().canUserEditTeam(text.split("/")[0]) ? (
-            <Link href={`../${text}`} target="_blank">
-              <Typography variant="body2">{text}</Typography>
-            </Link>
-          ) : (
-            <Typography variant="body2">{text}</Typography>
-          )
-        }
-        secondary={
-          <>
-            <Typography variant="body2" fontSize="small">
-              {nestedFlowLastPublishedTitle}
-            </Typography>
-            {summary && (
-              <Typography variant="body2" fontSize="small">
-                {summary}
-              </Typography>
-            )}
-          </>
-        }
-      />
-    </ListItem>
-  );
-};
-
-interface AlteredNodesSummary {
-  title: string;
-  portals: Portal[];
-  updated: number;
-  deleted: number;
-}
-
-const AlteredNodesSummaryContent = (props: {
-  alteredNodes: AlteredNode[];
-  url: string;
-}) => {
-  const { alteredNodes, url } = props;
-  const [expandNodes, setExpandNodes] = useState<boolean>(false);
-
-  const changeSummary: AlteredNodesSummary = {
-    title: "",
-    portals: [],
-    updated: 0,
-    deleted: 0,
-  };
-
-  alteredNodes.map((node) => {
-    if (node.id === "0") {
-      changeSummary["title"] =
-        "You are publishing the main service for the first time.";
-    } else if (node.id && Object.keys(node).length === 1) {
-      changeSummary["deleted"] += 1;
-    } else if (node.type === TYPES.InternalPortal) {
-      if (node.data?.text?.includes("/")) {
-        changeSummary["portals"].push({ ...node.data, flowId: node.id });
-      }
-    } else if (node.type) {
-      changeSummary["updated"] += 1;
-    }
-
-    return changeSummary;
-  });
-
-  return (
-    <Box pb={2}>
-      {changeSummary["title"] && (
-        <Typography variant="body2" pb={2}>
-          {changeSummary["title"]}
-        </Typography>
-      )}
-      {(changeSummary["updated"] > 0 || changeSummary["deleted"] > 0) && (
-        <Box pb={2}>
-          <ul>
-            <li key={"updated"}>
-              <Typography variant="body2">{`${changeSummary["updated"]} nodes have been updated or added`}</Typography>
-            </li>
-            <li key={"deleted"}>
-              <Typography variant="body2">{`${changeSummary["deleted"]} nodes have been deleted`}</Typography>
-            </li>
-          </ul>
-          <Box
-            style={{
-              display: "flex",
-              flexDirection: "column",
-              marginBottom: 2,
-            }}
-          >
-            <Box
-              style={{
-                display: "flex",
-                flexDirection: "row",
-                alignItems: "center",
-              }}
-            >
-              <Typography variant="body2">{`See detailed changelog `}</Typography>
-              <Button
-                onClick={() => setExpandNodes((expandNodes) => !expandNodes)}
-                size="small"
-                disableRipple
-              >
-                <Caret
-                  expanded={expandNodes}
-                  color="primary"
-                  titleAccess={
-                    expandNodes ? "Less information" : "More information"
-                  }
-                />
-              </Button>
-            </Box>
-            <Collapse in={expandNodes}>
-              <Box pb={1}>
-                <ul>
-                  {alteredNodes.map((node) => (
-                    <AlteredNodeListItem node={node} />
-                  ))}
-                </ul>
-              </Box>
-            </Collapse>
-          </Box>
-        </Box>
-      )}
-      <Divider />
-      {changeSummary["portals"].length > 0 && (
-        <Box pt={2}>
-          <Typography variant="body2">{`This includes recently published changes in the following nested services:`}</Typography>
-          <List sx={{ listStyleType: "disc", marginLeft: 4 }}>
-            {changeSummary["portals"].map((portal) => (
-              <AlteredNestedFlowListItem {...portal} />
-            ))}
-          </List>
-        </Box>
-      )}
-      <Divider />
-      <Box pt={2}>
-        <Typography variant="body2">
-          {`Preview these content changes in-service before publishing `}
-          <Link href={url.replace("/published", "/preview")} target="_blank">
-            {`here (opens in a new tab).`}
-          </Link>
-        </Typography>
-      </Box>
-    </Box>
-  );
-};
-
 const Sidebar: React.FC<{
   url: string;
 }> = React.memo((props) => {
@@ -385,7 +174,7 @@ const Sidebar: React.FC<{
   const [lastPublishedTitle, setLastPublishedTitle] = useState<string>(
     "This flow is not published yet",
   );
-  const [validationMessage, setValidationMessage] = useState<string>();
+  const [validationChecks, setValidationChecks] = useState<ValidationCheck[]>([]);
   const [alteredNodes, setAlteredNodes] = useState<AlteredNode[]>();
   const [dialogOpen, setDialogOpen] = useState<boolean>(false);
   const [summary, setSummary] = useState<string>();
@@ -393,6 +182,56 @@ const Sidebar: React.FC<{
 
   const handleChange = (event: React.SyntheticEvent, newValue: SidebarTabs) => {
     setActiveTab(newValue);
+  };
+
+  const handleCheckForChangesToPublish = async () => {
+    try {
+      setLastPublishedTitle("Checking for changes...");
+      const alteredFlow = await validateAndDiffFlow(flowId);
+      setAlteredNodes(
+        alteredFlow?.data.alteredNodes
+          ? alteredFlow.data.alteredNodes
+          : [],
+      );
+      setLastPublishedTitle(
+        alteredFlow?.data.alteredNodes
+          ? `Found changes to ${alteredFlow.data.alteredNodes.length} nodes`
+          : alteredFlow?.data.message,
+      );
+      setValidationChecks(alteredFlow?.data?.validationChecks);
+      setDialogOpen(true);
+    } catch (error) {
+      setLastPublishedTitle(
+        "Error checking for changes to publish",
+      );
+
+      if (error instanceof AxiosError) {
+        alert(error.response?.data?.error);
+      } else {
+        alert(
+          `Error checking for changes to publish. Confirm that your graph does not have any corrupted nodes and that all external portals are valid. \n${error}`,
+        );
+      }
+    }
+  };
+
+  const handlePublish = async () => {
+    try {
+      setDialogOpen(false);
+      setLastPublishedTitle("Publishing changes...");
+      const { alteredNodes, message } = await publishFlow(
+        flowId,
+        summary,
+      );
+      setLastPublishedTitle(
+        alteredNodes
+          ? `Successfully published changes to ${alteredNodes.length} nodes`
+          : `${message}` || "No new changes to publish",
+      );
+    } catch (error) {
+      setLastPublishedTitle("Error trying to publish");
+      alert(error);
+    }
   };
 
   const _lastPublishedRequest = useAsync(async () => {
@@ -521,36 +360,7 @@ const Sidebar: React.FC<{
                 variant="contained"
                 color="primary"
                 disabled={!useStore.getState().canUserEditTeam(teamSlug)}
-                onClick={async () => {
-                  try {
-                    setLastPublishedTitle("Checking for changes...");
-                    const alteredFlow = await validateAndDiffFlow(flowId);
-                    setAlteredNodes(
-                      alteredFlow?.data.alteredNodes
-                        ? alteredFlow.data.alteredNodes
-                        : [],
-                    );
-                    setLastPublishedTitle(
-                      alteredFlow?.data.alteredNodes
-                        ? `Found changes to ${alteredFlow.data.alteredNodes.length} node(s)`
-                        : alteredFlow?.data.message,
-                    );
-                    setValidationMessage(alteredFlow?.data.description);
-                    setDialogOpen(true);
-                  } catch (error) {
-                    setLastPublishedTitle(
-                      "Error checking for changes to publish",
-                    );
-
-                    if (error instanceof AxiosError) {
-                      alert(error.response?.data?.error);
-                    } else {
-                      alert(
-                        `Error checking for changes to publish. Confirm that your graph does not have any corrupted nodes and that all external portals are valid. \n${error}`,
-                      );
-                    }
-                  }
-                }}
+                onClick={handleCheckForChangesToPublish}
               >
                 CHECK FOR CHANGES TO PUBLISH
               </Button>
@@ -560,17 +370,26 @@ const Sidebar: React.FC<{
               onClose={() => setDialogOpen(false)}
               aria-labelledby="alert-dialog-title"
               aria-describedby="alert-dialog-description"
+              maxWidth="md"
             >
-              <DialogTitle style={{ paddingBottom: 0 }}>
-                {lastPublishedTitle}
+              <DialogTitle variant="h3" component="h1">
+                {`Check for changes to publish`}
               </DialogTitle>
               <DialogContent>
                 {alteredNodes?.length ? (
                   <>
                     <AlteredNodesSummaryContent
-                      alteredNodes={alteredNodes}
-                      url={props.url}
+                      alteredNodes={alteredNodes} lastPublishedTitle={lastPublishedTitle}
                     />
+                    <ValidationChecks validationChecks={validationChecks} />
+                    <Box pb={2}>
+                      <Typography variant="body2">
+                        {`Preview these content changes in-service before publishing `}
+                        <Link href={props.url.replace("/published", "/preview")} target="_blank">
+                          {`here (opens in a new tab).`}
+                        </Link>
+                      </Typography>
+                    </Box>
                     <Input
                       bordered
                       type="text"
@@ -580,38 +399,21 @@ const Sidebar: React.FC<{
                       onChange={(e) => setSummary(e.target.value)}
                     />
                   </>
-                ) : validationMessage ? (
-                  validationMessage
                 ) : (
-                  lastPublishedTitle
+                  <Typography variant="body2">
+                    {`No new changes to publish`}
+                  </Typography>
                 )}
               </DialogContent>
-              <DialogActions>
+              <DialogActions sx={{ paddingX: 2 }}>
                 <Button onClick={() => setDialogOpen(false)}>
                   KEEP EDITING
                 </Button>
                 <Button
                   color="primary"
-                  onClick={async () => {
-                    try {
-                      setDialogOpen(false);
-                      setLastPublishedTitle("Publishing changes...");
-                      const { alteredNodes, message } = await publishFlow(
-                        flowId,
-                        summary,
-                      );
-                      setLastPublishedTitle(
-                        alteredNodes
-                          ? `Successfully published changes to ${alteredNodes.length} node(s)`
-                          : `${message}` || "No new changes to publish",
-                      );
-                    } catch (error) {
-                      setLastPublishedTitle("Error trying to publish");
-                      alert(error);
-                      console.log(error);
-                    }
-                  }}
-                  disabled={!alteredNodes || alteredNodes.length === 0}
+                  variant="contained"
+                  onClick={handlePublish}
+                  disabled={!alteredNodes || alteredNodes.length === 0 || validationChecks.filter((v) => v.status === "Fail").length > 0}
                 >
                   PUBLISH
                 </Button>
